LayoutTests/canvas/philip/tests/size.attributes.setAttribute.hex-expected.txt

< [index] >
size.attributes.setAttribute.hex
Parsing of non-negative integers in setAttribute
References: size.nonnegativeinteger
Actual output:
Passed
